+++ docs/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Planned
 - TBD
 
+### Added
+- Optional security alert webhooks for abuse/rate-limit events
+
 ---
 
 ## [0.2.0] - 2026-01-19